chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
openai
